Create routers: auth, todos, clean app main.py file
