+++ todos_FastAPI/main.py
@@ -1,97 +1,20 @@
 """
-Add, remove, change, make prioritizations of todos.
-- create todos
+Authentication and authorization
+- create endpoint
 - make changes
-- delete todos
-- create a database for todos
+- create a database for user
 - validations
 """
-from typing import Annotated
+from fastapi import FastAPI
 
-from pydantic import BaseModel, Field
-from sqlalchemy.orm import Session
-
-from fastapi import FastAPI, HTTPException, Depends
-from fastapi.params import Path
-from starlette import status
-
-from models import Todos
-from database import engine, SessionLocal
-
-from routers import auth
+import models
+from database import engine
+from routers import auth, todos
 
 
 app = FastAPI()
-# Use if you do not have DB
+
 # models.Base.metadata.create_all(bind=engine)
-# Create a router
+
 app.include_router(auth.router)
-
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    except ConnectionError as e:
-        print(f"Ups, something went wrong: {e}")
-    finally:
-        db.close()
-
-
-db_dependency = Annotated[Session, Depends(get_db)]
-
-
-class TodoRequest(BaseModel):
-    title: str = Field(min_length=3)
-    description: str = Field(min_length=3, max_length=100)
-    priority: int = Field(gt=0, lt=6)
-    complete: bool
-
-
-@app.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
-
-
-@app.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-    if todo_model is not None:
-        return todo_model
-    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
-
-
-@app.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.dict())
-
-    db.add(todo_model)
-    db.commit()
-
-
-@app.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
-                      todo_request: TodoRequest,
-                      todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-    if todo_model is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
-
-    todo_model.title = todo_request.title
-    todo_model.description = todo_request.description
-    todo_model.priority = todo_request.priority
-    todo_model.complete = todo_request.complete
-
-    db.add(todo_model)
-    db.commit()
-
-
-@app.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency,
-                      todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-    if todo_model is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
-
-    db.query(Todos).filter(Todos.id == todo_id).delete()
-    db.commit()
+app.include_router(todos.router)
