+++ todos_FastAPI/routers/todos.py
@@ -1,0 +1,76 @@
+from typing import Annotated
+
+from sqlalchemy.orm import Session
+from fastapi import APIRouter, HTTPException, Depends
+from fastapi.params import Path
+from starlette import status
+
+from database import SessionLocal
+
+from models import Todos
+from validators import TodoRequest
+
+
+router = APIRouter()
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    except ConnectionError as e:
+        print(f"Ups, something went wrong: {e}")
+    finally:
+        db.close()
+
+
+db_dependency = Annotated[Session, Depends(get_db)]
+
+
+@router.get("/", status_code=status.HTTP_200_OK)
+async def read_all(db: db_dependency):
+    return db.query(Todos).all()
+
+
+@router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
+async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if todo_model is not None:
+        return todo_model
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
+
+
+@router.post("/todo", status_code=status.HTTP_201_CREATED)
+async def create_todo(db: db_dependency, todo_request: TodoRequest):
+    todo_model = Todos(**todo_request.dict())
+
+    db.add(todo_model)
+    db.commit()
+
+
+@router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def update_todo(db: db_dependency,
+                      todo_request: TodoRequest,
+                      todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if todo_model is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
+
+    todo_model.title = todo_request.title
+    todo_model.description = todo_request.description
+    todo_model.priority = todo_request.priority
+    todo_model.complete = todo_request.complete
+
+    db.add(todo_model)
+    db.commit()
+
+
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(db: db_dependency,
+                      todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if todo_model is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
+
+    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.commit()
