+++ todos_FastAPI/validators.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel, Field
+
+
+class TodoRequest(BaseModel):
+    title: str = Field(min_length=3)
+    description: str = Field(min_length=3, max_length=100)
+    priority: int = Field(gt=0, lt=6)
+    complete: bool
